Moved flight_xbox into xbox_controller
