+++ spacenavigator.py
@@ -7,122 +7,7 @@ from pywinusb.hid import usage_pages, helpers, winapi
 from djitellopy import Tello
 import cv2
 import easygui
-from inputs import get_gamepad
-import math
-import threading
-
-class XboxController(object):
-    MAX_TRIG_VAL = math.pow(2, 8)
-    MAX_JOY_VAL = math.pow(2, 15)
-
-    def __init__(self):
-
-        self.LeftJoystickY = 0
-        self.LeftJoystickX = 0
-        self.RightJoystickY = 0
-        self.RightJoystickX = 0
-        self.LeftTrigger = 0
-        self.RightTrigger = 0
-        self.LeftBumper = 0
-        self.RightBumper = 0
-        self.A = 0
-        self.X = 0
-        self.Y = 0
-        self.B = 0
-        self.LeftThumb = 0
-        self.RightThumb = 0
-        self.Back = 0
-        self.Start = 0
-        self.LeftDPad = 0
-        self.RightDPad = 0
-        self.UpDPad = 0
-        self.DownDPad = 0
-        # Dpad Fix
-        self.DPadY = 0
-        self.DPadX = 0
-
-        self._monitor_thread = threading.Thread(target=self._monitor_controller, args=())
-        self._monitor_thread.daemon = True
-        self._monitor_thread.start()
-
-
-    def read(self): # returns inputs
-        left_y = self.LeftJoystickX
-        left_x = self.LeftJoystickY
-        right_y = self.RightJoystickY
-        right_x = self.RightJoystickX
-        # Buttons
-        a = self.A
-        x = self.X
-        b = self.B
-        y = self.Y
-        # D pad
-        
-        Dpad_x = self.DPadX
-        Dpad_y = self.DPadY
-        # Bumbers
-        rb = self.RightBumper
-        rt = self.RightTrigger
-        lb = self.LeftBumper
-        lt = self.LeftTrigger
-        # Extras
-        start = self.Start
-        select = self.Back
-
-        # Extra for spacenavigator to controll lift
-        extra = self.RightTrigger - self.LeftTrigger
-        return [left_y, left_x, right_y, right_x, a, x, y, b, Dpad_x, Dpad_y, rb, rt, lb, lt, start, select, extra]
-
-
-    def _monitor_controller(self):
-        while True:
-            events = get_gamepad()
-            for event in events:
-                if event.code == 'ABS_Y':
-                    self.LeftJoystickY = event.state / XboxController.MAX_JOY_VAL # normalize between -1 and 1
-                elif event.code == 'ABS_X':
-                    self.LeftJoystickX = event.state / XboxController.MAX_JOY_VAL # normalize between -1 and 1
-                elif event.code == 'ABS_RY':
-                    self.RightJoystickY = event.state / XboxController.MAX_JOY_VAL # normalize between -1 and 1
-                elif event.code == 'ABS_RX':
-                    self.RightJoystickX = event.state / XboxController.MAX_JOY_VAL # normalize between -1 and 1
-                elif event.code == 'ABS_Z':
-                    self.LeftTrigger = event.state / XboxController.MAX_TRIG_VAL # normalize between 0 and 1
-                elif event.code == 'ABS_RZ':
-                    self.RightTrigger = event.state / XboxController.MAX_TRIG_VAL # normalize between 0 and 1
-                elif event.code == 'BTN_TL':
-                    self.LeftBumper = event.state
-                elif event.code == 'BTN_TR':
-                    self.RightBumper = event.state
-                elif event.code == 'BTN_SOUTH':
-                    self.A = event.state
-                elif event.code == 'BTN_NORTH':
-                    self.Y = event.state #previously switched with X
-                elif event.code == 'BTN_WEST':
-                    self.X = event.state #previously switched with Y
-                elif event.code == 'BTN_EAST':
-                    self.B = event.state
-                elif event.code == 'BTN_THUMBL':
-                    self.LeftThumb = event.state
-                elif event.code == 'BTN_THUMBR':
-                    self.RightThumb = event.state
-                elif event.code == 'BTN_SELECT':
-                    self.Back = event.state
-                elif event.code == 'BTN_START':
-                    self.Start = event.state
-                elif event.code == 'BTN_TRIGGER_HAPPY1':
-                    self.LeftDPad = event.state
-                elif event.code == 'BTN_TRIGGER_HAPPY2':
-                    self.RightDPad = event.state
-                elif event.code == 'BTN_TRIGGER_HAPPY3':
-                    self.UpDPad = event.state
-                elif event.code == 'BTN_TRIGGER_HAPPY4':
-                    self.DownDPad = event.state
-                # DPad Fix
-                elif event.code == 'ABS_HAT0X':
-                    self.DPadX = event.state
-                elif event.code == 'ABS_HAT0Y':
-                    self.DPadY = event.state
+from xbox_controller import *
 
 # current version number
 __version__ = "0.2.2"
@@ -533,9 +418,6 @@ def set_led(state):
 def flight(mode):
     tello = Tello()
     tello.connect()
-    print("Devices found:\n\t%s" % "\n\t".join(list_devices()))
-    dev = open(callback=None, button_callback=toggle_led)
-    help = 0
     # Connection the cam or not
     if mode == "no":    
         while True:
@@ -580,51 +462,12 @@ def flight(mode):
             elif a[7][0] == 1 and help != 0:
                 tello.flip("r")
             tello.send_rc_control(int(a[4]*100), int(a[5]*100), int(a[3]*100), int(a[6]*100))
-
-def flight_xbox():
-
-    tello = Tello()
-    tello.connect()
-    print("Devices found:\n\t%s" % "\n\t".join(list_devices()))
-    dev = open(callback=None, button_callback=toggle_led)
-    help = 0
-    cont = XboxController()
-    while True:
-
-        print(cont.read())
-        if cont.read()[15] == 1:
-            easygui.msgbox("Press 'Ok' to engage throw takeoff",title="Info")
-            tello.initiate_throw_takeoff()
-            help += 1
-        elif cont.read()[14] == 1 and help == 0:
-            tello.takeoff()
-            help += 1
-            print("Takeoff")
-        elif cont.read()[14] == 1 and help != 0:
-            tello.land()
-            break
-
-        elif cont.read()[9] == 1:
-            tello.flip("b")
-        
-        elif cont.read()[8] == -1:
-            tello.flip("l")
-
-        elif cont.read()[9] == -1:
-            tello.flip("f")
-
-        elif cont.read()[8] == 1:
-            tello.flip("r")
-        
-        tello.send_rc_control(int(cont.read()[0]*100), int(cont.read()[1]*100), int(cont.read()[16]*100), int(cont.read()[3]*100))
-
-      
+   
 
 if __name__ == "__main__":
     tello = Tello()
-    tello.connect()
-    # battery_temperature = tello.get_temperature()
-    # battery = tello.get_battery()
+    battery_temperature = tello.get_temperature()
+    battery = tello.get_battery()
 
     button_list = []
     button1 = "Video"
@@ -632,7 +475,7 @@ if __name__ == "__main__":
     button_list.append(button1)
     button_list.append(button2)
 
-    output = easygui.buttonbox(f"The current battery percentage is none%!\nPressing 'Ok' is starting the Live Video Feed", "Info", button_list)
+    output = easygui.buttonbox(f"The current battery percentage is {battery}%!\nPressing 'Ok' is starting the Live Video Feed", "Info", button_list)
     
     if output == "Video":
         tello.streamon()
@@ -642,36 +485,31 @@ if __name__ == "__main__":
     print("Devices found:\n\t%s" % "\n\t".join(list_devices()))
     dev = open(callback=None, button_callback=toggle_led)
     help = 0
-    cont = XboxController()
+
     while True:
 
+        if output == "Video":
+            img = tello.get_frame_read().frame
+            cv2.imshow("LiveStream", img)
+            cv2.waitKey(1)
 
-        print(cont.read())
-        if cont.read()[15] == 1:
+        a = read()
+        print(a)
+        sleep(0.1)
+        if a[7][1] == 1 and a[7][0] == 1:
             easygui.msgbox("Press 'Ok' to engage throw takeoff",title="Info")
             tello.initiate_throw_takeoff()
             help += 1
-        elif cont.read()[14] == 1 and help == 0:
+        elif a[7][1] == 1 and help == 0:
             tello.takeoff()
             help += 1
             print("Takeoff")
-        elif cont.read()[14] == 1 and help != 0:
+        elif a[7][1] == 1 and help != 0:
             tello.land()
             break
-
-        elif cont.read()[9] == 1:
-            tello.flip("b")
-        
-        elif cont.read()[8] == -1:
-            tello.flip("l")
-
-        elif cont.read()[9] == -1:
-            tello.flip("f")
-
-        elif cont.read()[8] == 1:
+        elif a[7][0] == 1 and help != 0:
             tello.flip("r")
-        
-        tello.send_rc_control(int(cont.read()[0]*100), int(cont.read()[1]*100), int(cont.read()[16]*100), int(cont.read()[3]*100))
+        tello.send_rc_control(int(a[4]*100), int(a[5]*100), int(a[3]*100), int(a[6]*100))
         
     flight_time = tello.get_flight_time()
     easygui.msgbox(f"Total flight time was: {flight_time} seconds", title="Info-")
